+++ spring-cloud-tencent-rpc-enhancement/src/main/java/com/tencent/cloud/rpc/enhancement/config/RpcEnhancementAutoConfiguration.java
@@ -188,6 +188,7 @@ public class RpcEnhancementAutoConfiguration {
 		public PolarisLoadBalancerRequestTransformer polarisLoadBalancerRequestTransformer() {
 			return new PolarisLoadBalancerRequestTransformer();
 		}
+
 	}
 
 	/**