+++ CHANGELOG.md
@@ -15,3 +15,4 @@
 - build(deps): bump guava in /spring-cloud-tencent-dependencies
 - fix:fix reporting bug when port is -1.
 - fix:update guava version.
+- fix:fix circuit breaker bean load order bug when using Nacos discovery.